Moves ReaderState to esc
